@@ -25,9 +25,12 @@
     "@types/jest": "^25.2.1",
     "@types/lodash": "^4.14.202",
     "@types/node": "^13.13.5",
-    "@types/react-router-dom": "5.1.7",
     "@typescript-eslint/eslint-plugin": "^5.48.2",
     "@typescript-eslint/parser": "^5.48.2",
+    "@types/react-router-dom": "5.3.3",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.0",
+    "@types/react-redux": "7.1.30",
     "cspell": "^8.6.1",
     "eslint": "^8.32.0",
     "eslint-config-airbnb": "^18.2.1",
@@ -42,7 +45,8 @@
     "jest": "26.6.0",
     "lint-staged": ">=10",
     "prettier": "3.0.3",
-    "turbo": "1.10.15"
+    "turbo": "1.10.15",
+    "react-scripts": "5.0.1"
   },
   "dependencies": {
     "@dfinity/agent": "1.0.1",
@@ -69,8 +73,11 @@
     "dayjs": "^1.11.10",
     "jsbi": "^4.3.0",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-router-dom": "5.2.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "5.3.4",
+    "redux": "4.2.1",
+    "react-redux": "8.0.5",
     "tiny-invariant": "^1.3.1"
   },
   "author": "",
