@@ -17,6 +17,9 @@ import GlobalSteps from "components/Steps/index";
 import ActorInitial from "components/Actor";
 import { GlobalContext } from "hooks/useGlobalContext";
 import TransactionsUpdater from "store/transactions/updater";
+import { WagmiProvider } from "wagmi";
+import { wagmiConfig } from "constants/wagmi";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import Web3Provider from "./components/Web3Injector";
 import { useFetchICPPrices, useFetchAllSwapTokens } from "./store/global/hooks";
@@ -45,35 +48,41 @@ export default function App() {
 
   useFetchSnsAllTokensInfo();
 
+  const queryClient = new QueryClient();
+
   return (
-    <StyledEngineProvider injectFirst>
-      <Web3Provider>
-        <Route component={GoogleAnalytics} />
-        <TransactionsUpdater />
-        <ThemeProvider theme={theme(customization)}>
-          <SnackbarProvider maxSnack={100}>
-            <GlobalContext.Provider value={{ AllPools }}>
-              <ActorInitial>
-                <CssBaseline />
-                <NavigationScroll>
-                  {isInitialStandardLoading ? (
-                    <Loader />
-                  ) : (
-                    <ErrorBoundary>
-                      <Routes />
-                    </ErrorBoundary>
-                  )}
-                  <Snackbar />
-                  <FullscreenLoading />
-                  <GlobalSteps />
-                  {isConnected ? <RiskStatement /> : null}
-                  {walletConnectorOpen ? <WalletConnector /> : null}
-                </NavigationScroll>
-              </ActorInitial>
-            </GlobalContext.Provider>
-          </SnackbarProvider>
-        </ThemeProvider>
-      </Web3Provider>
-    </StyledEngineProvider>
+    <WagmiProvider config={wagmiConfig}>
+      <QueryClientProvider client={queryClient}>
+        <StyledEngineProvider injectFirst>
+          <Web3Provider>
+            <Route component={GoogleAnalytics} />
+            <TransactionsUpdater />
+            <ThemeProvider theme={theme(customization)}>
+              <SnackbarProvider maxSnack={100}>
+                <GlobalContext.Provider value={{ AllPools }}>
+                  <ActorInitial>
+                    <CssBaseline />
+                    <NavigationScroll>
+                      {isInitialStandardLoading ? (
+                        <Loader />
+                      ) : (
+                        <ErrorBoundary>
+                          <Routes />
+                        </ErrorBoundary>
+                      )}
+                      <Snackbar />
+                      <FullscreenLoading />
+                      <GlobalSteps />
+                      {isConnected ? <RiskStatement /> : null}
+                      {walletConnectorOpen ? <WalletConnector /> : null}
+                    </NavigationScroll>
+                  </ActorInitial>
+                </GlobalContext.Provider>
+              </SnackbarProvider>
+            </ThemeProvider>
+          </Web3Provider>
+        </StyledEngineProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
   );
 }
